webapp/requirements/base.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
Django==4.0.3
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-environ==0.8.1
mysqlclient==2.1.0
soupsieve==2.3.1
sqlparse==0.4.2
